@@ -1,5 +1,6 @@
-﻿using Newtonsoft.Json;
-using System.Runtime.Serialization;
+﻿using System.Runtime.Serialization;
+using System.Text.Json;
+using System.Text.Json.Serialization;
 
 namespace Seq.App.aspsms.Models
 {
@@ -12,7 +13,12 @@ namespace Seq.App.aspsms.Models
 
         public static Result GetFromJson(string json)
         {
-            return JsonConvert.DeserializeObject<Result>(json);
+            return JsonSerializer.Deserialize<Result>(json, new JsonSerializerOptions { NumberHandling = JsonNumberHandling.AllowReadingFromString });
+        }
+
+        public string GetAsJson()
+        {
+            return JsonSerializer.Serialize(this, new JsonSerializerOptions { DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull, NumberHandling = JsonNumberHandling.WriteAsString });
         }
     }
 }
