@@ -1,32 +1,33 @@
-﻿using Seq.App.aspsms.Models;
-using Xunit;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Seq.App.aspsms.Models;
 
 namespace Seq.App.aspsms.Tests
 {
+    [TestClass]
     public class ResultTests
     {
-        [Fact]
+        [TestMethod]
         public void CheckValues()
         {
             // Act
             var data = Result.GetFromJson("{\"StatusCode\":\"1\",\"StatusInfo\":\"OK\"}");
 
             // Assert
-            Assert.Equal("1", data.StatusCode);
-            Assert.Equal("OK", data.StatusInfo);
-            Assert.Null(data.Credits);
+            Assert.AreEqual("1", data.StatusCode);
+            Assert.AreEqual("OK", data.StatusInfo);
+            Assert.IsNull(data.Credits);
         }
 
-        [Fact]
+        [TestMethod]
         public void CheckValuesWithCredits()
         {
             // Act
             var data = Result.GetFromJson("{\"StatusCode\":\"1\",\"StatusInfo\":\"OK\",\"Credits\":\"567\"}");
 
             // Assert
-            Assert.Equal("1", data.StatusCode);
-            Assert.Equal("OK", data.StatusInfo);
-            Assert.Equal(567, data.Credits);
+            Assert.AreEqual("1", data.StatusCode);
+            Assert.AreEqual("OK", data.StatusInfo);
+            Assert.AreEqual(567, data.Credits);
         }
     }
 }
