@@ -1,6 +1,8 @@
-﻿namespace Seq.App.aspsms.Models
+﻿using System.Text.Json;
+
+namespace Seq.App.aspsms.Models
 {
-    public class CheckCredits : JsonBase
+    public class CheckCredits
     {
         public string UserName { get; private set; }
         public string Password { get; private set; }
@@ -9,6 +11,11 @@
         {
             UserName = userName;
             Password = password;
+        }
+
+        public string GetAsJson()
+        {
+            return JsonSerializer.Serialize(this, new JsonSerializerOptions { DefaultIgnoreCondition = System.Text.Json.Serialization.JsonIgnoreCondition.WhenWritingNull });
         }
     }
 }
